refactor: extract vxtwitter feature
